# app/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample API for address searching and geocoding using Dadata API.",
        "title": "My API",
        "termsOfService": "http://localhost:8080/swagger/index.html",
        "contact": {},
        "version": "1.1"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
        "/address/geocode": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "geocode"
                ],
                "summary": "Geocode an address",
                "operationId": "geocodeAddress",
                "parameters": [
                    {
                        "description": "Geocode Request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.GeocodeRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.GeocodeResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/address/search": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "search"
                ],
                "summary": "Search for an address",
                "operationId": "searchAddress",
                "parameters": [
                    {
                        "description": "Search Request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.SearchRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.SearchResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/login": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "SingIn"
                ],
                "summary": "Log in a user",
                "operationId": "loginUser",
                "parameters": [
                    {
                        "description": "User",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.User"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "JWT token"
                    },
                    "401": {
                        "description": "Invalid credentials"
                    }
                }
            }
        },
        "/register": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "SingUp"
                ],
                "summary": "Register a new user",
                "operationId": "registerUser",
                "parameters": [
                    {
                        "description": "User",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.User"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "User registered successfully"
                    },
                    "400": {
                        "description": "Invalid request format"
                    }
                }
            }
        }
    },
    "definitions": {
        "main.AddressGeo": {
            "description": "AddressGeo represents the geocode result for an address.",
            "type": "object",
            "properties": {
                "suggestions": {
                    "description": "The list of suggestions for the geocode result\n@jsonField suggestions\n@example Example suggestions: [{\"value\": \"123 Main St\", \"unrestricted_value\": \"123 Main St, City\", \"data\": {\"postal_code\": \"12345\", \"country\": \"USA\"}}]",
                    "type": "array",
                    "items": {
                        "type": "object",
                        "properties": {
                            "data": {
                                "type": "object",
                                "properties": {
                                    "country": {
                                        "type": "string"
                                    },
                                    "postal_code": {
                                        "type": "string"
                                    }
                                }
                            },
                            "unrestricted_value": {
                                "type": "string"
                            },
                            "value": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "main.AddressSearch": {
            "description": "AddressSearch represents the search result for an address.",
            "type": "object",
            "properties": {
                "metro": {
                    "description": "The list of metro stations nearby\n@jsonField metro\n@example Example metro stations: [{\"distance\": 1.5, \"line\": \"A\", \"name\": \"Station 1\"}]",
                    "type": "array",
                    "items": {
                        "type": "object",
                        "properties": {
                            "distance": {
                                "type": "number"
                            },
                            "line": {
                                "type": "string"
                            },
                            "name": {
                                "type": "string"
                            }
                        }
                    }
                },
                "result": {
                    "description": "The result address\n@jsonField result\n@example 123 Main St",
                    "type": "string"
                },
                "source": {
                    "description": "The source of the address information\n@jsonField source\n@example Dadata",
                    "type": "string"
                }
            }
        },
        "main.GeocodeRequest": {
            "description": "GeocodeRequest represents the request body for address geocoding.",
            "type": "object",
            "properties": {
                "lat": {
                    "description": "The latitude\n@jsonQuery query\n@example Example latitude: \"40.7128\"",
                    "type": "string"
                },
                "lon": {
                    "description": "The longitude\n@jsonQuery query\n@example Example longitude: \"-74.0060\"",
                    "type": "string"
                }
            }
        },
        "main.GeocodeResponse": {
            "description": "GeocodeResponse represents the response body for address geocoding.",
            "type": "object",
            "properties": {
                "addressesGeo": {
                    "description": "The geocode result for the given coordinates\n@jsonField addressesGeo\n@example Example response:\n{\"suggestions\": [{\"value\": \"123 Main St\", \"unrestricted_value\": \"123 Main St, City\", \"data\": {\"postal_code\": \"12345\", \"country\": \"USA\"}}]}",
                    "allOf": [
                        {
                            "$ref": "#/definitions/main.AddressGeo"
                        }
                    ]
                }
            }
        },
        "main.SearchRequest": {
            "description": "SearchRequest represents the request body for address search",
            "type": "object",
            "properties": {
                "query": {
                    "description": "The search query\n@jsonQuery query\n@example Example search query: \"123 Main St\"",
                    "type": "string"
                }
            }
        },
        "main.SearchResponse": {
            "description": "SearchResponse represents the response body for address search.",
            "type": "object",
            "properties": {
                "addressesSearch": {
                    "description": "The list of addresses matching the search query\n@jsonField addressesSearch\n@example Example response:\n[{\"source\": \"Dadata\", \"result\": \"123 Main St\", \"metro\": []}]",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/main.AddressSearch"
                    }
                }
            }
        },
        "main.User": {
            "type": "object",
            "properties": {
                "login": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}